Correcciones en el conteo del carrito de compras (habia un error, no contaba bien) SOLUCIONADO
